move more to util
