+++ ee/clickhouse/queries/util.py
@@ -1,6 +1,8 @@
 from datetime import datetime, timedelta
-from typing import Optional, Tuple
+from typing import Any, Dict, Optional, Tuple
 
+from ee.clickhouse.client import ch_client
+from ee.clickhouse.sql.events import GET_EARLIEST_TIMESTAMP_SQL
 from posthog.models.filter import Filter
 
 
@@ -19,3 +21,34 @@ def parse_timestamps(filter: Filter) -> Tuple[Optional[str], Optional[str]]:
     date_to = "and timestamp < '{}'".format(_date_to.strftime("%Y-%m-%d 00:00:00"))
 
     return date_from, date_to
+
+
+def get_time_diff(interval: str, start_time: Optional[datetime], end_time: Optional[datetime]) -> Tuple[int, int]:
+
+    _start_time = start_time or ch_client.execute(GET_EARLIEST_TIMESTAMP_SQL)[0][0]
+    _end_time = end_time or datetime.now()
+
+    time_diffs: Dict[str, Any] = {
+        "minute": 60,
+        "hour": 3600,
+        "day": 3600 * 24,
+        "week": 3600 * 24 * 7,
+        "month": 3600 * 24 * 30,
+    }
+
+    diff = _end_time - _start_time
+    return int(diff.total_seconds() / time_diffs[interval]), time_diffs[interval]
+
+
+def get_interval_annotation(interval: Optional[str]) -> str:
+    if interval is None:
+        return "toStartOfDay"
+
+    map: Dict[str, str] = {
+        "minute": "toStartOfMinute",
+        "hour": "toStartOfHour",
+        "day": "toStartOfDay",
+        "week": "toStartOfWeek",
+        "month": "toStartOfMonth",
+    }
+    return map[interval]
